Add a placeholder secrets.env
I didn't realize docker-compose failed if secrets.env was missing.
So I've added a placeholder with no variables to prevent this from
being a stumbling block to new developers.

The 'secrets.env' file will remain listed in '.gitignore' so that
updates to the file won't be accidentally commited to the repository.